Move hydrostatic_dz to vertical_coordinates.py
