+++ external/vcm/vcm/calc/vertical_coordinates.py
@@ -1,7 +1,7 @@
 import numpy as np
 import xarray as xr
 from ..cubedsphere.constants import COORD_Z_CENTER, COORD_Z_OUTER
-from .constants import _GRAVITY
+from .constants import _GRAVITY, _RVGAS, _RDGAS
 
 _TOA_PRESSURE = 300.0  # Pa
 _REVERSE = slice(None, None, -1)
@@ -182,3 +182,23 @@ def surface_pressure_from_delp(
     )
 
     return surface_pressure
+
+
+def hydrostatic_dz(
+    T: xr.DataArray, q: xr.DataArray, delp: xr.DataArray, dim: str = COORD_Z_CENTER
+) -> xr.DataArray:
+    """Compute layer thickness assuming hydrostatic balance.
+
+    Args:
+        T: temperature
+        q: specific humidity
+        delp: pressure thickness
+        dim (optional): name of vertical dimension. Defaults to "pfull".
+
+    Returns:
+        layer thicknesses dz
+    """
+    pi = pressure_at_interface(delp, dim_center=dim, dim_outer=dim)
+    tv = T * (1 + (_RVGAS / _RDGAS - 1) * q)
+    dlogp = np.log(pi).diff(dim)
+    return -dlogp * _RDGAS * tv / _GRAVITY
